refactor(home): extract grid toolbar + search bar (stage 2, slice 4)
Moves the two large inline subtrees out of CharacterCardGrid.build into
focused widgets under lib/ui/pages/home/widgets/:
- HomeGridToolbar — selection/organize header or folder breadcrumb or the
  mode toggle, plus the sort dropdown, grid-size slider, and the
  refresh/multi-select/organize/new-folder/import actions. The
  _getActiveFolderName helper moved with it.
- HomeGridSearchBar — the search field + folder-scope prefix selector.

Subtrees extracted verbatim; the widgets take exactly the fields each
used. The two mode bottom-bars stay inline (small).

character_card_grid.dart: 776 → 437 — now under the 500-line cap. analyze
clean; the "3 characters" home golden (which renders the toolbar/search)
shows only the pre-existing 27px environment-AA diff, identical to the
untouched baseline — no rendering change.
